feat : component-separation / optimized-structure
