+++ api/__init__.py
@@ -1,0 +1,9 @@
+"""
+AutoPromptix API - Presentation Layer
+
+REST API server for AutoPromptix core functionality.
+"""
+
+from .server import APIServer
+
+__all__ = ['APIServer']
