--- docker/Dockerfile.simple
@@ -1,39 +1,0 @@
-# Simple AutoPromptix Docker Image
-FROM python:3.11-slim
-
-# Set environment variables
-ENV PYTHONUNBUFFERED=1
-ENV PYTHONDONTWRITEBYTECODE=1
-ENV BUILD_FRONTEND=false
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    curl \
-    git \
-    && rm -rf /var/lib/apt/lists/*
-
-# Set work directory
-WORKDIR /app
-
-# Copy requirements first for better caching
-COPY requirements.txt .
-
-# Install Python dependencies
-RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r requirements.txt
-
-# Copy application code
-COPY . .
-
-# Frontend build skipped - using built-in HTML dashboard
-
-# Create non-root user
-RUN useradd --create-home --shell /bin/bash autopromptix && \
-    chown -R autopromptix:autopromptix /app
-USER autopromptix
-
-# Expose ports
-EXPOSE 8000 8001
-
-# Default command
-CMD ["python", "enhanced_main.py"]
